update sidecar with SR rest proxy
